requirements.txt
flask
pypdf
python-dotenv
google-genai
openai
faiss-cpu
numpy
